Added support for moai breakpoints and callback debugging.
This required specific logic for path comparison in the breakpoint check
as the moai host is using full path where the standard lua interpreter is
using a relative path the debugger expects.

Updated logic in modebug.on() call to correctly set debug hook on the
current thread (even when it is a main thread). This is probably how it
should have been done anyway, but also allows debugging of moai callbacks
registered with setCallback.