Added conversion of file names on windows to lower case.
This is needed because some users may have folders in different
capitalizations between the file system and their spelling in
package.path. For example, if there is a folder 'foo', but in
the code it is referenced as package.path = [[c:\FOO\?.lua]], Lua still
loads the file bar.lua from folder 'foo' successfully (as the Windows
treats 'FOO' and 'foo' names as being the same). However, it will
then report in the debugger the name as c:\FOO\bar.lua (as specified
in package.path, even though the "real" name in the file system is
C:\foo\bar.lua. This mismatch creates problems with breakpoints in the
debugger, so to avoid it the file names are converted to lower case (on
Windows only).