Performance improvements (thanks to Stephen Nichols for profiling and providing detailed suggestions).
The most effect should be felt when the debugging was started using
start() call and proceeded without stepping through the code at least once
(in_debugger function continued to be called in this case, which was not
necessary). Removed creating and throwing away temporary tables in
select() call as this was causing noticeable slowdown in tests.