Improved path matching when absolute and relative paths are used.
The path used in breakpoints is always as short as possible with basedir
being removed from it if possible. This makes a difference when an
application is started with a full name (lua /full/path.lua) in the folder
'/full', but then loads another file from the same folder, which will have
a relative name in the debugger. In this case, the basedir (/full in this
case) will be removed from the absolute path, allowing breakpoints to
work.